I slept for a few hours, but this was giving me nightmares, so it's fixed. As it turns out, Javascript does have faculties to be not-terrible.
To be more precice, I finally deduced that someone was using a for..in.. loop in an array,
which apparently iterated over prototype'd additions as well, that is, unless those additions
have the [[Enumeratable]] tag or somesuch set to false. And, as it would happen, Javascript
does, in fact, have the ability to set the internal attributes of a property as you define it,
meaning it is possible to safely extend a prototype without messing up people's (for .. in .. )
loops.
Oh, also, sorry. I was kinda missing from the code-style talk (and the github setup talk) because
of class, so didn't get the memo on code style specs until, like, now, when I saw github had a
wiki.
Probably useful for the people working on the overseer/AI controller, I also learned of the methods
Object.seal(o) and Object.freeze(o), which could be useful for protecting your AIs/overseers/whatever
from malicious code. They mess with the attributes of the object's properties and lock it down.